# backend/src/main/java/com/travelassistant/trip/api/CreateTripRequest.java
package com.travelassistant.trip.api;
import jakarta.validation.Valid; import jakarta.validation.constraints.*; import java.time.LocalDate; import java.util.List;
public record CreateTripRequest(@NotBlank @Size(max=200) String destination,@NotNull LocalDate startDate,@Min(1)@Max(30) int days,@NotNull@Valid MoneyRequest budget,@Min(1)@Max(50) int travelers,@NotNull@Size(max=20) List<@NotBlank@Size(max=50) String> preferences,@NotBlank@Size(max=64) String timezone,@Size(max=2000) String additionalRequirements) {}